fix restaurant detail
